chore: split segment to another file
